lemonlot-backend/src/main/java/com/revature/thelemonlot/util/JwtUtil.java
package com.revature.thelemonlot.util;

import java.security.Key;
import java.util.Base64;
import java.util.Date;

import org.springframework.beans.factory.annotation.Value;
import org.springframework.stereotype.Component;

import com.revature.thelemonlot.model.User;

import io.jsonwebtoken.Claims;
import io.jsonwebtoken.ExpiredJwtException;
import io.jsonwebtoken.Jwts;
import io.jsonwebtoken.MalformedJwtException;
import io.jsonwebtoken.SignatureAlgorithm;
import io.jsonwebtoken.UnsupportedJwtException;
import io.jsonwebtoken.io.Decoders;
import io.jsonwebtoken.security.Keys;
import io.jsonwebtoken.security.SignatureException;
import io.jsonwebtoken.security.WeakKeyException;

@Component
public class JwtUtil {

    @Value("${SECRET_KEY}")
    private String secretKey;
    private final long validityInMilliseconds = 24 * 60 * 60 * 1000; // 1 day in milliseconds

    public String generateToken(String subject, String role) {
        if (role == null || role.isEmpty())
            role = "USER";
        Date now = new Date();
        Date expiryDate = new Date(now.getTime() + validityInMilliseconds);

        Key signingKey = Keys.hmacShaKeyFor(Decoders.BASE64.decode(secretKey));

        return Jwts.builder()
                .setSubject(subject)
                .setIssuedAt(now)
                .setExpiration(expiryDate)
                .claim("role", role)
                .signWith(signingKey, SignatureAlgorithm.HS256)
                .compact();
    }

    public Claims decodeJWT(String token) {
        try {
            return Jwts.parserBuilder()
                    .setSigningKey(Keys.hmacShaKeyFor(Base64.getDecoder().decode(secretKey)))
                    .build()
                    .parseClaimsJws(token)
                    .getBody();
        } catch (ExpiredJwtException | MalformedJwtException | UnsupportedJwtException | SignatureException | WeakKeyException | IllegalArgumentException e) {
            throw new RuntimeException("Invalid JWT token. " + e.getMessage());
        }
    }

    public String extractUsername(String token) {
        return decodeJWT(token).getSubject();
    }

    public String extractRole(String token) {
        Claims claims = decodeJWT(token);
        return claims.get("role", String.class); // Default to null if the role is not present
    }

    public boolean validateToken(String token, User user) {
        if (token == null || user == null)
            return false;
        final String username = extractUsername(token);
        if (username == null || user.getUsername() == null)
            return false;
        return username.equals(user.getUsername()) && !isTokenExpired(token);
    }

    private boolean isTokenExpired(String token) {
        return decodeJWT(token).getExpiration().before(new Date());
    }
}